pixelformats: Use pixel groups instead of verticalsubsampling
At some point we have horizontalsubsampling and verticalsubsampling.
Then we had bytespergroup and verticalsubsampling, i.e. a mix of the
concept of pixel group and subsampling.

Let's just drop the subsampling totally, as we can represent the
subsamplings with a rectangular pixel group.

So, now we have 'pixelspergroup' for each PixelFormat, which tells the
width and height, in pixels and lines, of a pixel group. And for each
plane, we have 'bytespergroup' which tells the number of bytes a pixel
group takes horizontally on that plane, and 'linespergroup' which tells
the number of lines a pixel group takes on that plane.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>