conv: Support 16 bit debayering
When bayer format variants are sent for demosaic, the data can overflow
a 16 bit dtype, which sums a 3x3 window of values. Extend this local
data to use np.uint32 unconditionally as the data type to ensure the
addition of multiple 16 bit values will not overflow.

The existing SRGGB10, SRGGB12, and SRGGB16 test data was generated when
the demosaic functions were overflowing. This caused the reference data
to be invalid as it contains the overflowed values.

While SRGGB10, and SRGGB12 pixels can not overflow a 16 bit calculation
of 9 summed values; due to the test data generator simply generating
random data, the input data has values larger than 10 and 12 bits, which
feeds directly into the calculation. As a result, these are essentially
calculated as if they were 16 bit pixel values.

To ensure the tests continue to pass, generate new data, which has been
visually inspected not to overflow.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>