tests: test_conv: Switch to dynamic testing
Drop the reference images, and just generate the data on the fly,
relying on sha256 checks. This way we don't need to store large files
in the repository, files that can be trivially generated.

For convenience, the test_conv.py can be ran with "--save" to save the
frame files.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ideasonboard.com>